chore(sidebar): move load sidebar width effect to sidebar hook
